Fixed bug where too many photon paths could overflow counter

+++ source/core/rendering/rendering_driver_finalframe.cpp
@@ -92,7 +92,7 @@ void Driver_finalframe::bake_photons(uint32_t frame) noexcept {
 
     auto const start = std::chrono::high_resolution_clock::now();
 
-    uint32_t num_paths = 0;
+    uint64_t num_paths = 0;
     uint32_t begin     = 0;
 
     float const iteration_threshold = photon_settings_.iteration_threshold;
@@ -109,7 +109,7 @@ void Driver_finalframe::bake_photons(uint32_t frame) noexcept {
             static_cast<int32_t>(begin), static_cast<int32_t>(photon_settings_.num_photons));
 
         for (uint32_t i = 0, len = thread_pool_.num_threads(); i < len; ++i) {
-            num_paths += photon_infos_[i].num_paths;
+            num_paths += static_cast<uint64_t>(photon_infos_[i].num_paths);
         }
 
         if (!num_paths) {
